packaging/windows/requirements.txt
PyYAML
cexprtk
customtkinter
imageio
numpy
openimageio
pillow
platformdirs
